# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ilevel API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{category_id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by id",
                "description": "Get category by id with all products",
                "operationId": "App\\Http\\Controllers\\API\\CategoryController::show",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category by id",
                "description": "Update category by id with all relations",
                "operationId": "App\\Http\\Controllers\\API\\CategoryController::update",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer lbsiiejroiehfuwoefnelfiower2432420agfasdfsadfasdgrewgwergwergwergewrg",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category by id",
                "description": "Delete category by id with all relations",
                "operationId": "App\\Http\\Controllers\\API\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer lbsiiejroiehfuwoefnelfiower2432420agfasdfsadfasdgrewgwergwergwergewrg",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Create a new product",
                "operationId": "App\\Http\\Controllers\\API\\ProductController::store",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer lbsiiejroiehfuwoefnelfiower2432420agfasdfsadfasdgrewgwergwergwergewrg",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/products/{product_id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product by id",
                "description": "Update category by id with all relations",
                "operationId": "App\\Http\\Controllers\\API\\ProductController::update",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer lbsiiejroiehfuwoefnelfiower2432420agfasdfsadfasdgrewgwergwergwergewrg",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product by id",
                "description": "Delete product by id with all relations",
                "operationId": "App\\Http\\Controllers\\API\\ProductController::destroy",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Bearer lbsiiejroiehfuwoefnelfiower2432420agfasdfsadfasdgrewgwergwergwergewrg",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the category",
                        "type": "string",
                        "example": "Plans and Cars"
                    },
                    "products": {
                        "title": "Product",
                        "description": "Product category model",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Category"
                }
            },
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the product",
                        "type": "string",
                        "example": "Plan"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the product",
                        "type": "string",
                        "example": "This is product description"
                    },
                    "model": {
                        "title": "Model",
                        "description": "Model of the product",
                        "type": "string",
                        "example": "Model"
                    },
                    "categories": {
                        "title": "Category",
                        "description": "Product category model",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            },
...
            "StoreProductRequest": {
                "title": "Store Product request",
                "description": "Store Product request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "title": {
                        "title": "name",
                        "description": "Name of the new product",
                        "type": "string",
                        "example": "Plan"
                    },
                    "model": {
                        "title": "model",
                        "description": "Model of the new product",
                        "type": "string",
                        "example": "Model"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the new product",
                        "type": "string",
                        "example": "This is product description"
                    },
                    "categories": {
                        "title": "categories",
                        "description": "Categories id for new product",
                        "type": "string",
                        "example": "[1,2,3]"
                    }
                },
                "type": "object"
            },
            "UpdateCategoryRequest": {
                "title": "Update Category request",
                "description": "Update Category request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "title": {
                        "title": "name",
                        "description": "New name of the current category",
                        "type": "string",
                        "example": "Plans and Cars"
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "title": "Update Product request",
                "description": "Update Product request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "title": {
                        "title": "name",
                        "description": "name of the product",
                        "type": "string",
                        "example": "Plan"
                    },
                    "model": {
                        "title": "model",
                        "description": "model of the product",
                        "type": "string",
                        "example": "Model"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the product",
                        "type": "string",
                        "example": "This is product description"
                    },
                    "categories": {
                        "title": "categories",
                        "description": "Categories id for product",
                        "type": "string",
                        "example": "[1,2,3]"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}